refactor: rename authentication classes for consistency
Renamed `Auth*` classes to `Authentication*` for consistency.
Propagated changes project-wide.
